+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Unreleased
 ==========
 
+- Change `RoomPosition` and `Position` methods `look` and `look_for` to return a `Result` instead
+  of panicking when used in room not visible in the current tick (breaking)
+
 0.12.2 (2023-06-17)
 ===================